+++ framework/domain/src/test/java/is/codion/framework/domain/entity/EntityTest.java
@@ -23,6 +23,7 @@ import is.codion.framework.domain.TestDomain.Department;
 import is.codion.framework.domain.TestDomain.Detail;
 import is.codion.framework.domain.TestDomain.Employee;
 import is.codion.framework.domain.TestDomain.NoPk;
+import is.codion.framework.domain.entity.Entity.Key;
 import is.codion.framework.domain.entity.attribute.AttributeDefinition;
 
 import org.junit.jupiter.api.Test;
@@ -136,7 +137,7 @@ public final class EntityTest {
 		dept1.put(Department.ID, 3);
 		dept2.put(Department.ID, 4);
 
-		Collection<Entity.Key> originalPrimaryKeys = Entity.originalPrimaryKeys(asList(dept1, dept2));
+		Collection<Key> originalPrimaryKeys = Entity.originalPrimaryKeys(asList(dept1, dept2));
 		assertTrue(originalPrimaryKeys.contains(entities.primaryKey(Department.TYPE, 1)));
 		assertTrue(originalPrimaryKeys.contains(entities.primaryKey(Department.TYPE, 2)));
 	}
@@ -149,7 +150,7 @@ public final class EntityTest {
 		Entity emp = entities.builder(Employee.TYPE)
 						.with(Employee.ID, 3)
 						.build();
-		Map<Entity.Key, Entity> entityMap = Entity.primaryKeyMap(asList(dept, emp));
+		Map<Key, Entity> entityMap = Entity.primaryKeyMap(asList(dept, emp));
 		assertSame(dept, entityMap.get(dept.primaryKey()));
 		assertSame(emp, entityMap.get(emp.primaryKey()));
 
@@ -161,10 +162,10 @@ public final class EntityTest {
 
 	@Test
 	void groupKeysByType() {
-		Entity.Key dept = entities.primaryKey(Department.TYPE, 1);
-		Entity.Key emp = entities.primaryKey(Employee.TYPE, 3);
+		Key dept = entities.primaryKey(Department.TYPE, 1);
+		Key emp = entities.primaryKey(Employee.TYPE, 3);
 
-		LinkedHashMap<EntityType, List<Entity.Key>> mapped = Entity.groupKeysByType(asList(dept, emp));
+		LinkedHashMap<EntityType, List<Key>> mapped = Key.groupByType(asList(dept, emp));
 		assertEquals(dept, mapped.get(Department.TYPE).get(0));
 		assertEquals(emp, mapped.get(Employee.TYPE).get(0));
 	}
@@ -273,7 +274,7 @@ public final class EntityTest {
 		Entity emp4 = entities.builder(Employee.TYPE)
 						.build();
 
-		Collection<Entity.Key> referencedKeys = Entity.keys(Employee.DEPARTMENT_FK, asList(emp1, emp2, emp3, emp4));
+		Collection<Key> referencedKeys = Entity.keys(Employee.DEPARTMENT_FK, asList(emp1, emp2, emp3, emp4));
 		assertEquals(2, referencedKeys.size());
 		referencedKeys.forEach(key -> assertEquals(Department.TYPE, key.entityType()));
 		Collection<Integer> values = Entity.values(new ArrayList<>(referencedKeys));
@@ -289,7 +290,7 @@ public final class EntityTest {
 						.with(NoPk.COL2, 2)
 						.with(NoPk.COL3, 3)
 						.build();
-		Collection<Entity.Key> keys = Entity.primaryKeys(singletonList(noPk));
+		Collection<Key> keys = Entity.primaryKeys(singletonList(noPk));
 		assertFalse(keys.iterator().next().isNull());
 		assertFalse(keys.iterator().next().primaryKey());
 	}